api/requirements.txt
fastapi
sqlalchemy
uvicorn
celery
flower
redis
pydantic
python-multipart
pyyaml
redislite
